# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Signer Service
    description: Signer service exposes HTTP API for making and verifying digital signatures and proofs for Verifiable Credentials.
    version: ""
host: localhost:8085
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /v1/credential:
        post:
            tags:
                - signer
            summary: CreateCredential signer
            description: CreateCredential creates VC with proof from raw JSON data.
            operationId: signer#CreateCredential
            parameters:
                - name: x-origin
                  in: header
                  description: Origin of call
                  required: true
                  type: string
                - name: CreateCredentialRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/SignerCreateCredentialRequestBody'
                    required:
                        - namespace
                        - key
                        - credentialSubject
                        - group
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: string
                        format: binary
            schemes:
                - http
    /v1/credential/proof:
        post:
            tags:
                - signer
            summary: CredentialProof signer
            description: CredentialProof adds a proof to a given Verifiable Credential.
            operationId: signer#CredentialProof
            parameters:
                - name: CredentialProofRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/SignerCredentialProofRequestBody'
                    required:
                        - namespace
                        - key
                        - credential
                        - group
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: string
                        format: binary
            schemes:
                - http
    /v1/credential/verify:
        post:
            tags:
                - signer
            summary: VerifyCredential signer
            description: VerifyCredential verifies the proof of a Verifiable Credential.
            operationId: signer#VerifyCredential
            parameters:
                - name: x-format
                  in: header
                  description: format
                  required: false
                  type: string
                  default: ldp_vc
                  enum:
                    - ldp_vc
                    - vc+sd-jwt
                - name: x-namespace
                  in: header
                  description: Namespace for status check
                  required: false
                  type: string
                - name: x-group
                  in: header
                  description: Group for status check
                  required: false
                  type: string
                - name: VerifyCredentialRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/SignerVerifyCredentialRequestBody'
                    required:
                        - credential
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/SignerVerifyCredentialResponseBody'
                        required:
                            - valid
            schemes:
                - http
    /v1/did/configuration:
        get:
            tags:
                - signer
            summary: DidConfiguration signer
            description: Creates did configuration.
            operationId: signer#DidConfiguration
            parameters:
                - name: x-namespace
                  in: header
                  description: Namespace.
                  required: true
                  type: string
                - name: x-group
                  in: header
                  description: Group.
                  required: false
                  type: string
                  default: ""
                - name: x-origin
                  in: header
                  description: origin for the configuration.
                  required: true
                  type: string
                - name: x-did
                  in: header
                  description: did for the Configuration
                  required: false
                  type: string
                - name: x-nonce
                  in: header
                  description: Proof challenge
                  required: false
                  type: string
                - name: DidConfigurationRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/SignerDidConfigurationRequestBody'
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: string
                        format: binary
            schemes:
                - http
...
    /v1/did/list:
        get:
            tags:
                - signer
            summary: didList signer
            operationId: signer#didList
            parameters:
                - name: x-namespace
                  in: header
                  description: Namespace.
                  required: true
                  type: string
                - name: x-group
                  in: header
                  description: Group.
                  required: false
                  type: string
                  default: ""
                - name: x-engine
                  in: header
                  description: Engine which shall be used. Eg transit or transit;kv or kv
                  required: false
                  type: string
                  default: transit
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/SignerDidListOKResponseBody'
                        required:
                            - list
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/SignerDidListNotFoundResponseBody'
                        required:
                            - list
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/SignerDidListInternalServerErrorResponseBody'
                        required:
                            - list
            schemes:
                - http
    /v1/jwk/{namespace}/{key}:
        get:
            tags:
                - signer
            summary: JwkPublicKey signer
            description: JwkPublicKey returns public key by name and namespace.
            operationId: signer#JwkPublicKey
            parameters:
                - name: namespace
                  in: path
                  description: Key namespace.
                  required: true
                  type: string
                - name: key
                  in: path
                  description: Key name.
                  required: true
                  type: string
                - name: JwkPublicKeyRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/SignerJwkPublicKeyRequestBody'
                    required:
                        - group
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: string
                        format: binary
                "404":
                    description: Not Found response.
                    schema:
                        type: string
                        format: binary
                "500":
                    description: Internal Server Error response.
                    schema:
                        type: string
                        format: binary
            schemes:
                - http
...
    /v1/namespaces/{namespace}/keys:
        get:
            tags:
                - signer
            summary: NamespaceKeys signer
            description: NamespaceKeys returns all keys in a given namespace.
            operationId: signer#NamespaceKeys
            parameters:
                - name: namespace
                  in: path
                  description: Namespace for signing keys.
                  required: true
                  type: string
                - name: x-group
                  in: header
                  description: Group identifies a group of keys within a namespace.
                  required: false
                  type: string
                  default: ""
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: array
                        items:
                            type: string
                            example: Voluptatibus aperiam eius velit.
            schemes:
                - http
    /v1/presentation:
        post:
            tags:
                - signer
            summary: CreatePresentation signer
            description: CreatePresentation creates VP with proof from raw JSON data.
            operationId: signer#CreatePresentation
            parameters:
                - name: CreatePresentationRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/SignerCreatePresentationRequestBody'
                    required:
                        - namespace
                        - key
                        - data
                        - group
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: string
                        format: binary
            schemes:
                - http
...
definitions:
...
    SignerDidConfigurationRequestBody:
        title: SignerDidConfigurationRequestBody
        type: object
        properties:
            x-signatureType:
                type: string
                description: Type of signature
                default: jsonwebsignature2020
                example: ed25519signature2020
                enum:
                    - ed25519signature2020
                    - jsonwebsignature2020
        example:
            x-signatureType: ed25519signature2020
...
    SignerJwkPublicKeyRequestBody:
        title: SignerJwkPublicKeyRequestBody
        type: object
        properties:
            engine:
                type: string
                description: Engine which shall be used. Eg transit or transit;kv or kv
                default: transit
                example: transit
            group:
                type: string
                description: Group identifies a group of keys within a namespace.
                default: ""
                example: Group
        example:
            engine: transit
            group: Group
        required:
            - group
...
    SignerVerifyCredentialResponseBody:
        title: SignerVerifyCredentialResponseBody
        type: object
        properties:
            disclosedCredential:
                type: string
                description: Disclosed Credential
                example: Quo voluptatem aut quisquam unde blanditiis.
                format: binary
            valid:
                type: boolean
                description: Valid specifies if the proof is successfully verified.
                example: false
        example:
            disclosedCredential: Unde consequatur odit.
            valid: false
        required:
            - valid
...